Weather_Info_Scraper/requirements.txt
requests~=2.32.3
python-dotenv~=1.1.0